ai-services/health-core/assessment/unified-health-assessment.service.ts
import { Injectable } from '@nestjs/common';
import { Logger } from '../../shared/utils/logger';
import { CacheManager } from '../../shared/utils/cache-manager';
import { DataProcessor } from '../../shared/utils/data-processor';
import { AIModelManager } from '../../shared/utils/ai-model-manager';
import { HealthData } from '../../shared/types/health.types';
import { IAssessmentResult } from '../../shared/types/assessment.types';

@Injectable()
export class UnifiedHealthAssessmentService {
  private readonly logger = new Logger(UnifiedHealthAssessmentService.name);

  constructor(
    private readonly cacheManager: CacheManager,
    private readonly dataProcessor: DataProcessor,
    private readonly modelManager: AIModelManager
  ) {}

  /**
   * 评估生命体征数据
   */
  async assessVitalSigns(healthData: HealthData): Promise<IAssessmentResult> {
    try {
      // 生成缓存键
      const cacheKey = `vital-signs-${healthData.userId}-${Date.now()}`;

      // 检查缓存
      const cachedResult = await this.cacheManager.get<IAssessmentResult>(cacheKey);
      if (cachedResult) {
        this.logger.info('使用缓存的生命体征评估结果', { userId: healthData.userId });
        return cachedResult;
      }

      // 数据处理
      const processedData = await this.dataProcessor.processHealthData(healthData);

      // 使用AI模型评估
      const modelResult = await this.modelManager.predict('vital-signs', processedData);

      // 处理评估结果
      const result: IAssessmentResult = {
        userId: healthData.userId,
        timestamp: new Date(),
        assessmentType: 'vital-signs',
        scores: {
          overall: modelResult.overallScore,
          categories: {
            bloodPressure: modelResult.categoryScores.bloodPressure,
            heartRate: modelResult.categoryScores.heartRate,
            bodyTemperature: modelResult.categoryScores.bodyTemperature,
            bloodOxygen: modelResult.categoryScores.bloodOxygen
          }
        },
        recommendations: modelResult.recommendations,
        alerts: modelResult.alerts,
        metadata: {
          modelVersion: modelResult.modelVersion,
          confidenceScores: modelResult.confidenceScores,
          dataQuality: processedData.metadata
        }
      };

      // 缓存结果
      await this.cacheManager.set(cacheKey, result, 3600); // 缓存1小时
      
      this.logger.info('完成健康评估', { userId: healthData.userId });
      return result;
    } catch (error) {
        this.logger.error('生命体征评估失败', error);
        throw error;
    }
  }

  /**
   * 评估生活方式数据
   */
  async assessLifestyle(healthData: HealthData): Promise<IAssessmentResult> {
    try {
      // 生成缓存键
      const cacheKey = `lifestyle-${healthData.userId}-${Date.now()}`;

      // 检查缓存
      const cachedResult = await this.cacheManager.get<IAssessmentResult>(cacheKey);
      if (cachedResult) {
        this.logger.info('使用缓存的生活方式评估结果', { userId: healthData.userId });
        return cachedResult;
      }

      // 数据处理
      const processedData = await this.dataProcessor.processHealthData(healthData);

      // 分析生活模式
      const patterns = this.analyzePatterns(healthData);

      // 生成建议
      const recommendations = this.generateRecommendations(patterns);

      // 使用AI模型评估
      const modelResult = await this.modelManager.predict('lifestyle', {
        data: processedData,
        patterns,
        recommendations
      });

      // 处理评估结果
      const result: IAssessmentResult = {
        userId: healthData.userId,
        timestamp: new Date(),
        assessmentType: 'lifestyle',
        scores: {
          overall: modelResult.overallScore,
          categories: {
            sleep: modelResult.categoryScores.sleep,
            activity: modelResult.categoryScores.activity,
            nutrition: modelResult.categoryScores.nutrition,
            stress: modelResult.categoryScores.stress
          }
        },
        patterns,
        recommendations: modelResult.recommendations,
        alerts: modelResult.alerts,
        metadata: {
          modelVersion: modelResult.modelVersion,
          confidenceScores: modelResult.confidenceScores,
          dataQuality: processedData.metadata
        }
      };

      // 缓存结果
      await this.cacheManager.set(cacheKey, result, 3600);

      this.logger.info('完成生活方式评估', { userId: healthData.userId });
      return result;
    } catch (error) {
      this.logger.error('生活方式评估失败', error);
      throw error;
    }
  }

  /**
   * 分析生活模式
   */
  private analyzePatterns(data: HealthData): any {
    // TODO: 实现生活模式分析逻辑
    return {};
  }

  /**
   * 生成建议
   */
  private generateRecommendations(patterns: any): any {
    // TODO: 实现建议生成逻辑
    return {};
  }
} 